editted doctor postuplenya

+++ app/models/doctors.py
@@ -139,6 +139,10 @@ class DoctorMonthlyPlan(Base):
             if user_plan.current_amount < 0:
                 raise HTTPException(status_code=404, detail="Med rep plan should be grater than 0 for tis product")
             if self.monthly_plan == 0:
+                result = await db.execute(select(DoctorPostupleniyaFact).filter(DoctorPostupleniyaFact.doctor_id==self.doctor_id, DoctorPostupleniyaFact.product_id==self.product_id))
+                postupleniya = result.scalars().first()
+                if postupleniya:
+                    raise HTTPException(status_code=400, detail="There is postuplenuya fact whith this product in this doctor")
                 query = f"delete from doctor_monthly_plan WHERE id={self.id}"
                 result = await db.execute(text(query))
             await db.commit()
